+++ sdk/monitor/azure-monitor-opentelemetry/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Features Added
 
+- Enable live metrics feature
+    ([#35218](https://github.com/Azure/azure-sdk-for-python/pull/35218))
+
 ### Breaking Changes
 
 ### Bugs Fixed